+++ tests/message-response-map.test.js
@@ -12,6 +12,7 @@ import { IMPORT_BACKGROUND_ACTIONS } from "../src/background/import-action-handl
 import { EXECUTION_TELEMETRY_ACTIONS } from "../src/background/telemetry-action-handler.ts";
 import { UPDATE_BACKGROUND_ACTIONS } from "../src/background/update-action-handler.ts";
 import { SYNC_BACKGROUND_ACTIONS } from "../src/background/sync-action-handler.ts";
+import { BACKUP_BACKGROUND_ACTIONS } from "../src/background/backup-action-handler.ts";
 import { GM_AUDIO_ACTIONS } from "../src/background/gm-audio-handler.ts";
 import { GM_COOKIE_ACTIONS } from "../src/background/gm-cookie-handler.ts";
 import { GM_MENU_ACTIONS } from "../src/background/gm-menu-handler.ts";
@@ -27,6 +28,7 @@ const ROUTED_BACKGROUND_ACTIONS = [
   ...EXECUTION_TELEMETRY_ACTIONS,
   ...UPDATE_BACKGROUND_ACTIONS,
   ...SYNC_BACKGROUND_ACTIONS,
+  ...BACKUP_BACKGROUND_ACTIONS,
   ...GM_AUDIO_ACTIONS,
   ...GM_COOKIE_ACTIONS,
   ...GM_MENU_ACTIONS,